Track last clicked issue
This is step 1 in a multi step process to increase the usefulness and decrease the annoying factor of codetriage. We can start by keeping track of which issues people actually click.

We'll expose this information in the subject line when a user hasn't clicked on an email in a few days.

Eventually we could use this information to send less emails. For instance if a user hasn't clicked on an email for the last 3 days, wait ~7 days before sending them another. We can always tweak this logic.

Either way the beginning point is tracking user clicks, and exposing that information

cc/ @olivierlacan