Keyset pagination on Repo#show
Offset pagination has the nasty habit of being REALLY slow when your offset is large. We don’t need random access for issues and doc methods, honestly having them on this page I don’t even think are that useful, but it looks nice.

This reduces the load on the DB from offset pagination and also removes the need to count all records on every page load. 

These counts add a non-trivial amount of DB load:

```
493:04:18.903353 | 10.4%          | 101,625,003 | 52:09:48.599802  | SELECT COUNT(*) FROM "issues" WHERE "issues"."repo_id" = $1 AND "issues"."state" = $2
```