Prompt users to add badge on repo creation
When a user adds a new repo, send them information on how to add a CodeTriage badge. Many do not realize the exist. Also add information on how to get more people contributing to their repo.

This is based on two pieces of feedback from  maintainers. One is “how can I get more people to subscribe to help out with my project” and the second was “I had no idea code triage had badges”. This will help without being obnoxious to all users.

In the future we could perhaps see if someone subscribing to a repo has commit on that repo and give them a similar message instead of only the first person to add the repo.