Cache the description meta tag
Before this change, we were doing two queries per request to get the
count of users and repos to generate the description meta tag.
Presumably, we don't actually care that these numbers are super
up-to-date, so we can cache it up to an hour (same as the similar text
in the template's h2 tag).