Change “days since clicked” to “streaks”
The current method “days since you clicked” is a bad incentive. The number grows while you do the wrong thing and it gets set back to zero when you do the right thing (click an issue). Instead we move to positive reinforcement. When the user clicks an issue they see the count increment by one. When the miss clicking on an email, the count is decremented (so if someone takes a short break they don’t lose their entire streak, but if they go away entirely it will.

In the future we can add some more semantic context around the number. For instance we might give the user positive reinforcement when `emails_missed_since_click` is 0 (since that means they are active recently), and some negative reinforcement if this value is higher (indicating that their streak is going down). We might also want to encourage someone to decrease the rate of their emails if their streak count is falling rather than them being not-engaged or deleting their account.

Other ideas instead of simple increment/decrement are some kind of “levels” where people can hit achievements. We could also perhaps add badges. 

The best way to understand the logic is to look at the `user_test.rb`. Essentially the `raw_streak_count` is the last recorded streak value. Every time we send an email we increment the `raw_emails_since_click`. With these two numbers we can calculate the `effective_streak_count` which is the value we show to the user. 

When a user clicks on a link then the `raw_streak_count` is set to the current effective streak count and then incremented (since they just clicked). We then reset the `raw_emails_since_click` counter to zero.

In the future it might make sense to introduce some kind of an “email record” table to record what issues and docs went out on which email, but for now all we need is to add these two numbers and we get a good amount of information.