Give more context around email settings
In addition to the raw result of “this setting does this action” it helps to have a mental framework for what the user is desiring to happen. If they think they will be a core member only responding to one issue a month (max) then their expectations aren’t realistic. If that’s not their goal, then it’s totally cool to get fewer emails. 

We want make sure that people’s goals are aligned with their settings.