Test DOC parsing and move to run in a fork
Parsing DOCs is extremely expensive on memory and our workers are constantly swapping. The goal is to mitigate this a tad by doing all the known memory intensive work inside of a fork. After we're done the fork goes away and with it, all of it's memory.

If this doesn't help then the next step would be to patch YARD to allow for the ability to incrementally generate docs instead of generating all the docs and loading into memory up front.