Fewer SQL queries when assigning issues
Right now we execute a SQL query for every repo subscribed to multiplied by the number of issues a person wants to receive for that repo.

With this PR we now try to find all of the issues with one query, this reduces the number of queries when sending out issues to the number of repos subscribed to rather than the number of repos subscribed to multiplied by the number of issues.