Add context to Scout for background jobs
Scout allows "context" to be added to background jobs https://docs.scoutapm.com/?_ga=2.6482226.1047178103.1572110586-117176759.1572110586#ruby-custom-context

This PR adds two classes, a UserBasedJob and a RepoBasedJob. When you inherit from a UserBasedJob then the user of that the job is processing on is automatically added to the scout context. Same for RepoBasedJob (but with repos instead of users).

I previously added code that allows these jobs to take either an ID or an object. I ported this code to live in the same callback that sets the context for scout.

I was chasing down some Redis memory OOM errors and trying to make smaller jobs. By using an integer instead of an object, we save on the serialization time for generating a globalID object and save for a small amount of space in Redis. It wouldn't be the end of the world if we ripped it out and ended up only using globalID and passing in objects, but since it's already in here and we're already using IDs for enqueues, we can keep this behavior.