Soft landing for non-post omniauth requests
Previously any request to `/users/<id>` would trigger a log in flow if you're not logged in. In https://github.com/omniauth/omniauth/pull/1010 (or somewhere around there) omniauth removed the ability to use a GET request to log in at all. Now all login requests must be done via GET. 

This means you cannot redirect to a login endpoint, you must drop the user off somewhere there is a form and they must manually click it to trigger a POST. 

That's what this PR does. If you try to edit or view your user account without being logged in today: it will give you a 404 error page with a cryptic error message.

```
Not found. Authentication passthru.
```

(Which I think omniauth could improve A LOT, FWIW). 

Anywhoo, to move forward I'm replacing the prior `authenticate_user!` before filter with one that checks if the request is a post. If it is, then it will directly authenticate the user. Otherwise we drop them on a "soft" page that tells them they must first login.

It's not perfect, I'm sure there will be edge cases, but it works.