Random via cached ID
The easiest way to work with random values is to call `order("RANDOM()")` but this is very bad for database performance. There's hacks to get faster lookups but they all come with caveats. Mostly in terms of guarantees, either they'll return no data, or less than the data you want. You've got to  retry queries, but if you do that in an unbounded way then you might end up an an infinite loop.

This approach attempts to give the maximum flexibility of full order("random()") but with non-awful performance. The cost here is that the values persist for the given time frame. Also since we're still using `order("RANDOM()")` the first query is still expensive.

It works like this: The SQL of the query is used to generate a cache key. The query is then made and the resulting IDs are returned. Then those IDs are used to query the original model.

I'm still not in love with it, but it's okay for now.

I think for more complex logic there's got to be a better way to handle this kind of processing. But 🤷‍♂️